Put UI update in IEnumerator. Update calculations.
With this update the UI only get's updated every second. Since the smallest amount of time is displayed in minutes, updating the UI every frame is wasted performance. Also updated the calculations a little to only calculate (countdown basically) when it is above zero (which is fair. Cant have a negative timer)